refactor(ui): extract utility functions from ViewBasedReportJobDetails
