+++ ui/apps/platform/src/Containers/Vulnerabilities/utils/vulnerabilityUtils.ts
@@ -2,6 +2,7 @@ import { severityRankings } from 'constants/vulnerabilities';
 import { isVulnerabilitySeverity } from 'types/cve.proto';
 import type { Exploit, VulnerabilitySeverity } from 'types/cve.proto';
 import type { NonEmptyArray } from 'utils/type.utils';
+import { getDate } from 'utils/dateUtils';
 
 /**
  * Get the highest severity of any vulnerability in the component array.
@@ -83,4 +84,42 @@ export function hasKnownExploit(exploit: Exploit | null | undefined): exploit is
 
 export function hasKnownRansomwareCampaignUse(exploit: Exploit | null | undefined) {
     return exploit?.knownRansomwareCampaignUse === 'Known';
+}
+
+/**
+ * Formats CVE discovered time filter values into user-friendly text
+ * @param value - Filter value like ">2024-01-01" or "2024-01-01"
+ * @returns Formatted string like "After January 1, 2024"
+ */
+export function formatCveDiscoveredTime(value: string): string {
+    try {
+        // Parse the condition prefix and date
+        const match = value.match(/^([<>]?)(.+)$/);
+        if (!match) {
+            return value; // Return original if parsing fails
+        }
+
+        const [, condition, dateStr] = match;
+        const date = new Date(dateStr);
+
+        // Check if date is valid
+        if (Number.isNaN(date.getTime())) {
+            return value; // Return original if date is invalid
+        }
+
+        const formattedDate = getDate(date);
+
+        // Map conditions to user-friendly text
+        switch (condition) {
+            case '>':
+                return `After ${formattedDate}`;
+            case '<':
+                return `Before ${formattedDate}`;
+            default:
+                return `On ${formattedDate}`;
+        }
+    } catch {
+        // Return original value if any error occurs
+        return value;
+    }
 }
