+++ ui/apps/platform/src/utils/arrayUtils.ts
@@ -13,3 +13,13 @@ export function toggleItemInArray<T>(
         : [...array, selectedItem];
     return newArray;
 }
+
+/**
+ * Normalizes a value to an array. If the value is already an array, returns it as-is.
+ * If it's a single value, wraps it in an array.
+ * @param value - The value to normalize to an array
+ * @returns An array containing the value(s)
+ */
+export function normalizeToArray<T>(value: T | T[]): T[] {
+    return Array.isArray(value) ? value : [value];
+}
